Refactor event package structure for better organization
Moved event classes to more specific subpackages within 'area/event' hierarchy. This enhances code maintainability by grouping related events together and improving overall project structure.
