+++ api/src/main/java/net/thenextlvl/protect/area/event/flag/AreaFlagResetEvent.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.protect.event;
+package net.thenextlvl.protect.area.event.flag;
 
 import lombok.Getter;
 import lombok.Setter;
