+++ api/src/main/java/net/thenextlvl/protect/area/event/member/AreaOwnerChangeEvent.java
@@ -1,10 +1,11 @@
-package net.thenextlvl.protect.event;
+package net.thenextlvl.protect.area.event.member;
 
 import core.annotation.ParametersAreNotNullByDefault;
 import core.annotation.TypesAreNotNullByDefault;
 import lombok.Getter;
 import lombok.Setter;
 import net.thenextlvl.protect.area.RegionizedArea;
+import net.thenextlvl.protect.area.event.AreaEvent;
 import org.bukkit.event.Cancellable;
 import org.jetbrains.annotations.Nullable;
