+++ api/src/main/java/net/thenextlvl/protect/area/event/player/PlayerAreaEvent.java
@@ -1,30 +1,34 @@
-package net.thenextlvl.protect.event;
+package net.thenextlvl.protect.area.event.player;
 
 import core.annotation.FieldsAreNotNullByDefault;
 import core.annotation.MethodsReturnNotNullByDefault;
 import core.annotation.ParametersAreNotNullByDefault;
 import core.annotation.TypesAreNotNullByDefault;
-import lombok.AccessLevel;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import net.thenextlvl.protect.area.Area;
-import org.bukkit.event.Event;
+import org.bukkit.entity.Player;
 import org.bukkit.event.HandlerList;
+import org.bukkit.event.player.PlayerEvent;
+
 
 /**
- * This class represents an abstract base class for events related to areas.
+ * This class represents an abstract base class for events related to areas and players.
  *
- * @param <T> The type of the area of this event
+ * @param <T> The type of the area associated with the event.
  */
 @Getter
 @TypesAreNotNullByDefault
 @FieldsAreNotNullByDefault
 @MethodsReturnNotNullByDefault
 @ParametersAreNotNullByDefault
-@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
-public abstract class AreaEvent<T extends Area> extends Event {
+public abstract class PlayerAreaEvent<T extends Area> extends PlayerEvent {
     private static final @Getter HandlerList handlerList = new HandlerList();
     private final T area;
+
+    public PlayerAreaEvent(Player player, T area) {
+        super(player);
+        this.area = area;
+    }
 
     @Override
     public HandlerList getHandlers() {
