+++ api/src/main/java/net/thenextlvl/protect/area/event/player/PlayerAreaLeaveEvent.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.protect.event;
+package net.thenextlvl.protect.area.event.player;
 
 import core.annotation.ParametersAreNotNullByDefault;
 import core.annotation.TypesAreNotNullByDefault;
