+++ plugin/src/main/java/net/thenextlvl/protect/area/CraftAreaService.java
@@ -8,10 +8,10 @@ import core.annotation.ParametersAreNotNullByDefault;
 import core.annotation.TypesAreNotNullByDefault;
 import lombok.RequiredArgsConstructor;
 import net.thenextlvl.protect.ProtectPlugin;
-import net.thenextlvl.protect.event.AreaCreateEvent;
-import net.thenextlvl.protect.event.AreaDeleteEvent;
-import net.thenextlvl.protect.event.PlayerAreaLeaveEvent;
-import net.thenextlvl.protect.event.PlayerAreaTransitionEvent;
+import net.thenextlvl.protect.area.event.AreaCreateEvent;
+import net.thenextlvl.protect.area.event.AreaDeleteEvent;
+import net.thenextlvl.protect.area.event.player.PlayerAreaLeaveEvent;
+import net.thenextlvl.protect.area.event.player.PlayerAreaTransitionEvent;
 import net.thenextlvl.protect.io.AreaAdapter;
 import org.bukkit.NamespacedKey;
 import org.bukkit.World;
