+++ plugin/src/main/java/net/thenextlvl/protect/area/CraftRegionizedArea.java
@@ -18,10 +18,10 @@ import core.annotation.TypesAreNotNullByDefault;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
-import net.thenextlvl.protect.event.AreaOwnerChangeEvent;
-import net.thenextlvl.protect.event.AreaRedefineEvent;
-import net.thenextlvl.protect.event.AreaSchematicDeleteEvent;
-import net.thenextlvl.protect.event.AreaSchematicLoadEvent;
+import net.thenextlvl.protect.area.event.member.AreaOwnerChangeEvent;
+import net.thenextlvl.protect.area.event.region.AreaRedefineEvent;
+import net.thenextlvl.protect.area.event.schematic.AreaSchematicDeleteEvent;
+import net.thenextlvl.protect.area.event.schematic.AreaSchematicLoadEvent;
 import net.thenextlvl.protect.area.event.member.AreaMemberAddEvent;
 import net.thenextlvl.protect.area.event.member.AreaMemberRemoveEvent;
 import org.bukkit.Location;
